rename: 배럴패턴적용, 이미지 업로더 파일명 변경
